test: split test file
Split the single test file to multiple files so it can be more manageable.
