+++ tests/test_redirect_core.py
@@ -1,162 +1,31 @@
 """
-Comprehensive test suite for the Memorial application.
+Core Redirect Functionality Tests
 
-This module tests the async-modernized Memorial application with 70+ tests covering:
-
-TEST CATEGORIES:
-  1. Favicon Endpoint Tests (4): Version handling, latest version, www compatibility, unconfigured sites
-  2. Site Image Endpoint Tests (5): Custom logos, local paths, defaults, hostname lookup, www normalization
-  3. WWW Normalization Tests (4): Config lookup normalization, double www, unconfigured behavior
-  4. Query Parameters Tests (4): Parameter preservation, special chars, empty values, path+query
-  5. Helper Function Tests (5): get_wayback_noframe_server_url(), get_host_configuration(), port stripping
-  6. URL Construction Tests (4): Wayback URLs with/without versions, noFrame preference, default selection
-  7. Edge Cases (11): Malformed headers, long paths, unicode, multiple slashes, language content, custom styling, timeouts, multi-site configs
-  8. Original Tests (~33): Status codes, metadata extraction, configuration, error handling
-
-KEY FEATURES TESTED:
-  - Async/ASGI Architecture: Async route handlers, async metadata extraction, HTTP timeout handling
-  - Configuration Management: Per-site overrides, environment variables, configuration precedence
-  - HTTP Features: Status codes, port normalization, WWW normalization, version timestamps
-  - Error Handling: Timeouts, malformed input, missing configuration fallbacks
-  - Content Customization: Metadata extraction control, static metadata, language-specific messages
-
-TEST PATTERNS:
-  - Mocking: unittest.mock.patch for config/dependencies, AsyncMock for async HTTP operations
-  - Assertions: Status codes, HTML content, configuration inheritance, URL construction
-  - Helpers: request_host(), get_title(), get_metadata(), with_test_config()
-
-RUNNING TESTS:
-  make test                    # Run all tests
-  make test-cov              # Run with coverage
-  pytest tests/test_basic.py -k favicon -v     # Run favicon tests
-  pytest tests/test_basic.py::test_name -v     # Run specific test
-
-STATISTICS:
-  Total Tests: 70 (original ~33 + 37 new)
-  Test Coverage: ~97% code coverage
-  Functional Areas: 7 major categories
+Core Memorial application tests including:
+  - Redirect routing
+  - Metadata extraction (with/without configuration)
+  - Configuration inheritance and precedence
+  - Status code handling
+  - Language-specific messages
+  - Custom templates
+  - Button styling
+  - Static content serving
 """
 
-from contextlib import contextmanager
-from unittest.mock import AsyncMock, Mock, patch
-
+import os
 import httpx
 import pytest
-from bs4 import BeautifulSoup
-import os
-
-# Import memorial first
+from unittest.mock import patch, AsyncMock, Mock
+from conftest import (
+    with_test_config, 
+    request_host, 
+    get_title, 
+    get_metadata,
+    mock_patcher,
+    mock_response,
+    client,
+)
 from memorial import app, fix_not_closed_metatags
-
-# Then patch memorial.httpx.AsyncClient to intercept calls from within memorial
-mock_patcher = patch("memorial.httpx.AsyncClient")
-mock_client_class = mock_patcher.start()
-
-# Create mock response that works for all tests
-mock_response = Mock()
-mock_response.status_code = 200
-mock_response.headers = {"content-type": "text/html"}
-mock_response.content = b"""
-<html>
-<head>
-    <title>Test Page</title>
-    <meta name="description" content="Test description">
-    <meta name="keywords" content="test, keywords">
-</head>
-<body>Test content</body>
-</html>
-"""
-
-# Configure the mock async client instance
-mock_client_instance = Mock()
-mock_client_instance.head = AsyncMock(return_value=mock_response)
-mock_client_instance.get = AsyncMock(return_value=mock_response)
-mock_client_instance.__aenter__ = AsyncMock(return_value=mock_client_instance)
-mock_client_instance.__aexit__ = AsyncMock(return_value=None)
-mock_client_class.return_value = mock_client_instance
-
-
-@pytest.fixture
-def client():
-    """Create a test client for the Quart app."""
-    return app.test_client()
-
-
-@contextmanager
-def with_test_config(archive_config):
-    """Context manager to temporarily set test configuration."""
-    # Save original state
-    if "ARCHIVE_CONFIG" in app.config:
-        original_archive_config = dict(app.config["ARCHIVE_CONFIG"])
-    else:
-        original_archive_config = {}
-
-    # Set test config
-    app.config["ARCHIVE_CONFIG"] = dict(archive_config)
-
-    # Ensure wayback URLs are set
-    if "WAYBACK_SERVER" not in app.config:
-        app.config["WAYBACK_SERVER"] = "https://arquivo.pt/wayback/"
-    if "WAYBACK_NOFRAME_SERVER" not in app.config:
-        app.config["WAYBACK_NOFRAME_SERVER"] = "https://arquivo.pt/noFrame/replay/"
-
-    try:
-        yield
-    finally:
-        app.config["ARCHIVE_CONFIG"] = original_archive_config
-
-
-async def request_host(client, path, host, expected_status=200):
-    """Helper method to make a request with a specific Host header.
-
-    Args:
-        client: The test client
-        path: The path to request
-        host: The Host header value
-        expected_status: Expected HTTP status code (default: 200)
-
-    Returns:
-        The response object
-    """
-    # Fake host so it properly match the template
-    response = await client.get(path, headers={"Host": host})
-    assert response.status_code == expected_status
-    return response
-
-
-def get_title(response_data):
-    """
-    Extract the title from the HTML content of the response.
-    """
-    html = str(response_data)
-    soup = BeautifulSoup(html, "html.parser")
-    title = soup.find("title")
-    if title:
-        print(f"Title: {title.text}")
-        return title.text
-    print("Title: None (not found)")
-    return None
-
-
-def get_metadata(response_data, meta_tag):
-    """
-    Extract the content of a specific meta tag from the HTML content of the response.
-
-    Args:
-        response_data: The HTML content of the response
-        meta_tag: The name of the meta tag to extract
-
-    Returns:
-        The content of the meta tag, or None if not found
-    """
-    html = str(response_data)
-    soup = BeautifulSoup(html, "html.parser")
-    meta = soup.find("meta", {"name": meta_tag})
-    if meta and "content" in meta.attrs:
-        print(f"{meta_tag}: {meta['content']}")
-        return meta["content"]
-    print(f"{meta_tag}: None (not found)")
-    return None
 
 
 @pytest.mark.asyncio
@@ -1054,763 +923,3 @@ async def test_strip_port_preserves_original_url(client):
 
 
 
-# Favicon Endpoint Tests: Verify /favicon.ico redirects to Arquivo.pt archived version
-# Tests: version handling, latest version, www compatibility, unconfigured sites
-# ==============================================================================
-# Tests for Favicon Endpoint (/favicon.ico)
-# ============================================================================== 
-
-
-@pytest.mark.asyncio
-async def test_favicon_redirect_with_version(client):
-    """
-    Test that favicon requests redirect to Arquivo.pt with the correct version.
-    """
-    with with_test_config({
-        "favicon-test.com": {
-            "version": "20200117175504",
-        }
-    }):
-        response = await client.get("/favicon.ico", headers={"Host": "favicon-test.com"})
-        assert response.status_code == 302  # Redirect status
-        location = response.headers.get("Location", "")
-        assert "arquivo.pt" in location
-        assert "20200117175504" in location
-        assert "favicon.ico" in location
-
-
-@pytest.mark.asyncio
-async def test_favicon_redirect_without_version(client):
-    """
-    Test that favicon requests redirect to latest version when no version specified.
-    """
-    with with_test_config({"favicon-noversion.com": {}}):
-        response = await client.get("/favicon.ico", headers={"Host": "favicon-noversion.com"})
-        assert response.status_code == 302
-        location = response.headers.get("Location", "")
-        assert "arquivo.pt" in location
-        assert "favicon-noversion.com" in location
-        assert "favicon.ico" in location
-
-
-@pytest.mark.asyncio
-async def test_favicon_with_www(client):
-    """
-    Test that favicon requests work correctly with www subdomain.
-    """
-    with with_test_config({"test-www.com": {"version": "20210101000000"}}):
-        response = await client.get("/favicon.ico", headers={"Host": "www.test-www.com"})
-        assert response.status_code == 302
-        location = response.headers.get("Location", "")
-        # Should normalize host to remove www
-        assert "test-www.com" in location
-
-
-@pytest.mark.asyncio
-async def test_favicon_unconfigured_site(client):
-    """
-    Test favicon redirect for unconfigured site.
-    """
-    response = await client.get("/favicon.ico", headers={"Host": "unconfigured-favicon.com"})
-    assert response.status_code == 302
-    location = response.headers.get("Location", "")
-    assert "arquivo.pt" in location
-
-
-
-
-# Site Image Endpoint Tests: Verify /memorial-site-image serves custom logos and defaults
-# Tests: custom URLs, local paths, hostname lookup, www normalization
-# ==============================================================================
-# Tests for Site Image Endpoint (/memorial-site-image)
-# ==============================================================================
-
-
-@pytest.mark.asyncio
-async def test_site_image_with_custom_logo_url(client):
-    """
-    Test that custom logo URL from config is used when provided.
-    """
-    with with_test_config({
-        "logo-test.com": {
-            "logo": "https://example.com/custom-logo.png"
-        }
-    }):
-        response = await client.get("/memorial-site-image", headers={"Host": "logo-test.com"})
-        # Response should be a redirect or served file
-        assert response.status_code in [200, 302, 404]
-
-
-@pytest.mark.asyncio
-async def test_site_image_with_local_logo_path(client):
-    """
-    Test that local logo path is correctly handled.
-    """
-    with patch("memorial.os.path.isdir") as mock_isdir:
-        mock_isdir.return_value = False  # Simulate images folder doesn't exist
-        with with_test_config({
-            "local-logo.com": {
-                "logo": "/static/img/custom-logo.png"
-            }
-        }):
-            response = await client.get("/memorial-site-image", headers={"Host": "local-logo.com"})
-            # Should attempt to serve the file
-            assert response.status_code in [200, 404]
-
-
-@pytest.mark.asyncio
-async def test_site_image_default_logo(client):
-    """
-    Test that default logo is used when no custom logo is configured.
-    """
-    with patch("memorial.os.path.isdir") as mock_isdir:
-        mock_isdir.return_value = False  # Simulate images folder doesn't exist
-        with with_test_config({"default-logo.com": {}}):
-            response = await client.get("/memorial-site-image", headers={"Host": "default-logo.com"})
-            # Should attempt to serve default logo
-            assert response.status_code in [200, 404]
-
-
-@pytest.mark.asyncio
-async def test_site_image_host_name_lookup(client):
-    """
-    Test that site images can be found by normalized host name (with dots replaced by underscores).
-    """
-    with patch("memorial.os.path.isdir") as mock_isdir:
-        with patch("memorial.os.listdir") as mock_listdir:
-            # Simulate finding a file matching the host name pattern
-            mock_isdir.return_value = True
-            mock_listdir.return_value = ["example_com.png", "other_file.jpg"]
-            
-            with with_test_config({"example.com": {}}):
-                response = await client.get("/memorial-site-image", headers={"Host": "example.com"})
-                # Should attempt to serve the file
-                assert response.status_code in [200, 404]
-
-
-@pytest.mark.asyncio
-async def test_site_image_with_www_normalization(client):
-    """
-    Test that www normalization works correctly in image lookup.
-    """
-    with patch("memorial.os.path.isdir") as mock_isdir:
-        with patch("memorial.os.listdir") as mock_listdir:
-            mock_isdir.return_value = True
-            mock_listdir.return_value = ["example_com.png"]
-            
-            with with_test_config({"example.com": {}}):
-                # Request with www should match example_com.png
-                response = await client.get("/memorial-site-image", headers={"Host": "www.example.com"})
-                assert response.status_code in [200, 404]
-
-
-@pytest.mark.asyncio
-async def test_site_image_folder_exists_finds_matching_file(client):
-    """Test line 262: os.path.isdir returns True and matching file is found.
-    
-    When images folder exists and contains a file matching the normalized hostname,
-    the function should find and attempt to serve it.
-    """
-    with patch("memorial.os.path.isdir") as mock_isdir:
-        with patch("memorial.os.listdir") as mock_listdir:
-            mock_isdir.return_value = True
-            # Return a list with files including the matching one
-            mock_listdir.return_value = ["readme.txt", "test_site_com.png", "other.jpg"]
-            
-            # Use host without config so it reaches the directory lookup
-            response = await client.get("/memorial-site-image", headers={"Host": "test-site.com"})
-            
-            # Verify os.path.isdir was called (covering line 262)
-            mock_isdir.assert_called()
-            # Verify os.listdir was called to search for matching files
-            mock_listdir.assert_called()
-
-
-@pytest.mark.asyncio
-async def test_site_image_folder_does_not_exist(client):
-    """Test line 262: os.path.isdir returns False.
-    
-    When images folder doesn't exist, the function should skip the directory
-    lookup and fall back to DEFAULT_LOGO.
-    """
-    with patch("memorial.os.path.isdir") as mock_isdir:
-        with patch("memorial.os.listdir") as mock_listdir:
-            # Line 262: if os.path.isdir(images_folder) returns False
-            mock_isdir.return_value = False
-            mock_listdir.return_value = []
-            
-            with patch.dict(
-                "memorial.app.config",
-                {
-                    "ARCHIVE_CONFIG": {"test-site.com": {}},
-                    "DEFAULT_LOGO": "default_logo.png",
-                    "IMAGES_FOLDER": "/nonexistent/folder",
-                }
-            ):
-                response = await client.get("/memorial-site-image", headers={"Host": "test-site.com"})
-                
-                # Verify os.path.isdir was called and returned False (line 262)
-                mock_isdir.assert_called()
-                # os.listdir should NOT be called when isdir is False
-                mock_listdir.assert_not_called()
-
-
-@pytest.mark.asyncio
-async def test_site_image_folder_exists_no_matching_files(client):
-    """Test line 262-263: os.path.isdir returns True but no matching files.
-    
-    When images folder exists but contains no files matching the hostname,
-    the function should fall back to DEFAULT_LOGO.
-    """
-    with patch("memorial.os.path.isdir") as mock_isdir:
-        with patch("memorial.os.listdir") as mock_listdir:
-            mock_isdir.return_value = True
-            # Return files that don't match the hostname
-            mock_listdir.return_value = ["other_site_com.png", "readme.txt", "config.json"]
-            
-            with patch.dict(
-                "memorial.app.config",
-                {
-                    "ARCHIVE_CONFIG": {"test-site.com": {}},
-                    "DEFAULT_LOGO": "default_logo.png",
-                    "IMAGES_FOLDER": "/static/img",
-                }
-            ):
-                response = await client.get("/memorial-site-image", headers={"Host": "test-site.com"})
-                
-                # Verify isdir was called and returned True (line 262)
-                mock_isdir.assert_called_with("/static/img")
-                # Verify listdir was called to search for matching files
-                mock_listdir.assert_called_with("/static/img")
-                # No matching file found, so DEFAULT_LOGO should be used
-
-
-@pytest.mark.asyncio
-async def test_site_image_listdir_raises_exception(client):
-    """Test line 262-268: os.listdir raises exception inside try block.
-    
-    When os.listdir raises an exception (permissions, I/O error, etc.),
-    the except block should catch it and fall back to DEFAULT_LOGO.
-    """
-    with patch("memorial.os.path.isdir") as mock_isdir:
-        with patch("memorial.os.listdir") as mock_listdir:
-            mock_isdir.return_value = True
-            # Raise an exception when listdir is called (line 263)
-            mock_listdir.side_effect = PermissionError("Access denied to images folder")
-            
-            with patch.dict(
-                "memorial.app.config",
-                {
-                    "ARCHIVE_CONFIG": {"test-site.com": {}},
-                    "DEFAULT_LOGO": "default_logo.png",
-                    "IMAGES_FOLDER": "/static/img",
-                }
-            ):
-                # Should not raise an exception, should gracefully fall back
-                response = await client.get("/memorial-site-image", headers={"Host": "test-site.com"})
-                
-                # Verify isdir returned True (line 262)
-                mock_isdir.assert_called_with("/static/img")
-                # Verify listdir was attempted and raised exception
-                mock_listdir.assert_called_with("/static/img")
-                # Should still return a response (with DEFAULT_LOGO)
-                assert response.status_code in [200, 404]
-
-
-@pytest.mark.asyncio
-async def test_site_image_real_filesystem_integration(client):
-    """Integration test line 262: Real filesystem check without mocks.
-    
-    This integration test creates a temporary directory and verifies that
-    os.path.isdir (line 262) is actually called and works with real paths,
-    not mocked operations. This ensures true code coverage of line 262.
-    """
-    import tempfile
-    import shutil
-    
-    # Create a temporary directory to serve as IMAGES_FOLDER
-    temp_dir = tempfile.mkdtemp()
-    try:
-        # Create a test image file matching the hostname pattern
-        test_image_path = os.path.join(temp_dir, "integration_test_com.png")
-        with open(test_image_path, "wb") as f:
-            f.write(b"PNG image data")
-        
-        with patch.dict(
-            "memorial.app.config",
-            {
-                "ARCHIVE_CONFIG": {"integration-test.com": {}},
-                "IMAGES_FOLDER": temp_dir,
-                "DEFAULT_LOGO": "default_logo.png",
-            }
-        ):
-            # Call without mocking os.path.isdir - uses real filesystem
-            response = await client.get("/memorial-site-image", headers={"Host": "integration-test.com"})
-            
-            # Verify the response (either serves the image or falls back to default)
-            assert response.status_code in [200, 404]
-            
-            # If 200, verify we got data
-            if response.status_code == 200:
-                data = await response.data
-                assert len(data) > 0
-    finally:
-        # Clean up temporary directory
-        shutil.rmtree(temp_dir)
-
-
-@pytest.mark.asyncio
-async def test_site_image_line_262_logo_with_images_folder_path(client):
-    """Test line 262: Logo containing images_folder path is split correctly.
-    
-    Line 262 executes: image_filename = logo.split(images_folder)[-1].lstrip("/\\")
-    This test ensures that when logo contains the images_folder path, it's
-    properly split to extract just the filename.
-    """
-    with patch.dict(
-        "memorial.app.config",
-        {
-            "ARCHIVE_CONFIG": {
-                "line262test.com": {
-                    "logo": "/static/img/custom-logo-262.png"  # Contains IMAGES_FOLDER
-                }
-            },
-            "IMAGES_FOLDER": "/static/img",
-            "DEFAULT_LOGO": "default.png",
-        }
-    ):
-        # Mock send_from_directory to verify it's called with the correct filename
-        with patch("memorial.send_from_directory") as mock_send:
-            mock_send.return_value = "Mocked response"
-            
-            response = await client.get("/memorial-site-image", headers={"Host": "line262test.com"})
-            
-            # Verify send_from_directory was called
-            mock_send.assert_called()
-            # Get the call arguments
-            call_args = mock_send.call_args
-            # Check that the filename was extracted correctly (line 262 execution)
-            # Should be "custom-logo-262.png" after split and lstrip
-
-
-# WWW Normalization Tests: Verify www. prefix is properly stripped for config lookup
-# Tests: config matching with www, double www handling, unconfigured behavior
-# ==============================================================================
-# Tests for WWW Normalization
-# ==============================================================================
-
-
-@pytest.mark.asyncio
-async def test_www_normalization_with_config(client):
-    """
-    Test that www is correctly stripped when looking up site configuration.
-    """
-    with with_test_config({"example.com": {"message_pt": "Site without www"}}):
-        response = await request_host(client, "/", "www.example.com", expected_status=200)
-        assert response.status_code == 200
-        html_content = (await response.data).decode("utf-8")
-        assert "Site without www" in html_content
-
-
-@pytest.mark.asyncio
-async def test_www_double_with_config(client):
-    """
-    Test that double www is also normalized correctly.
-    """
-    with with_test_config({"example.com": {"message_pt": "Configured site"}}):
-        # Even with www., should match example.com config
-        response = await request_host(client, "/", "www.example.com", expected_status=200)
-        assert response.status_code == 200
-        html_content = (await response.data).decode("utf-8")
-        assert "Configured site" in html_content
-
-
-@pytest.mark.asyncio
-async def test_www_without_config_returns_502(client):
-    """
-    Test that unconfigured site with www still returns 502.
-    """
-    with with_test_config({}):
-        response = await request_host(client, "/", "www.unconfigured-site.com", expected_status=502)
-        assert response.status_code == 502
-
-
-@pytest.mark.asyncio
-async def test_both_www_and_non_www_not_configured(client):
-    """
-    Test that if only non-www is configured, www requests match it.
-    """
-    with with_test_config({"target.com": {"status_code": 200}}):
-        response = await request_host(client, "/", "www.target.com", expected_status=200)
-        assert response.status_code == 200
-
-
-
-
-# Query Parameters Tests: Verify query string parameters are preserved in redirects
-# Tests: parameter preservation, special characters, empty values, path+query combinations
-# ==============================================================================
-# Tests for Query Parameters Handling
-# ==============================================================================
-
-
-@pytest.mark.asyncio
-async def test_query_parameters_preserved_in_response(client):
-    """
-    Test that query parameters are passed through to the template.
-    """
-    with with_test_config({"query-test.com": {"status_code": 200}}):
-        response = await request_host(client, "/?param1=value1&param2=value2", "query-test.com", expected_status=200)
-        assert response.status_code == 200
-        # The template should have access to query parameters via args
-
-
-@pytest.mark.asyncio
-async def test_query_parameters_with_special_characters(client):
-    """
-    Test that special characters in query parameters are handled correctly.
-    """
-    with with_test_config({"special-chars.com": {"status_code": 200}}):
-        response = await request_host(client, "/?search=hello%20world&filter=test%2Bvalue", "special-chars.com", expected_status=200)
-        assert response.status_code == 200
-
-
-@pytest.mark.asyncio
-async def test_empty_query_parameter(client):
-    """
-    Test handling of empty query parameters.
-    """
-    with with_test_config({"empty-param.com": {"status_code": 200}}):
-        response = await request_host(client, "/?empty=&filled=value", "empty-param.com", expected_status=200)
-        assert response.status_code == 200
-
-
-@pytest.mark.asyncio
-async def test_path_with_query_parameters(client):
-    """
-    Test that path and query parameters work correctly together.
-    """
-    with with_test_config({"path-query.com": {"status_code": 200}}):
-        response = await request_host(client, "/page/path?id=123&name=test", "path-query.com", expected_status=200)
-        assert response.status_code == 200
-
-
-
-
-# Helper Function Tests: Verify utility functions for host configuration and URL construction
-# Tests: get_wayback_noframe_server_url(), get_host_configuration(), port stripping
-# ==============================================================================
-# Tests for Helper Functions
-# ==============================================================================
-
-
-@pytest.mark.asyncio
-async def test_get_wayback_noframe_server_url_with_slash(client):
-    """
-    Test that get_wayback_noframe_server_url ensures trailing slash.
-    """
-    with patch.dict("memorial.app.config", {"WAYBACK_NOFRAME_SERVER": "https://arquivo.pt/noFrame/replay"}, clear=False):
-        from memorial import get_wayback_noframe_server_url
-        url = get_wayback_noframe_server_url()
-        assert url.endswith("/")
-        assert "arquivo.pt" in url
-
-
-@pytest.mark.asyncio
-async def test_get_wayback_noframe_server_url_custom(client):
-    """
-    Test that custom WAYBACK_NOFRAME_SERVER is used correctly.
-    """
-    custom_url = "https://custom.archive.org/replay/"
-    with patch.dict("memorial.app.config", {"WAYBACK_NOFRAME_SERVER": custom_url}, clear=False):
-        from memorial import get_wayback_noframe_server_url
-        url = get_wayback_noframe_server_url()
-        assert url == custom_url
-
-
-@pytest.mark.asyncio
-async def test_get_host_configuration_strips_www(client):
-    """
-    Test that get_host_configuration correctly normalizes hosts by removing www.
-    """
-    from memorial import get_host_configuration
-    
-    with patch.dict("memorial.app.config", {
-        "ARCHIVE_CONFIG": {"example.com": {"message_pt": "Test"}},
-        "STRIP_PORT": False
-    }, clear=False):
-        # Test with www
-        normalized_host, config = get_host_configuration("www.example.com")
-        assert normalized_host == "example.com"
-        assert config is not None
-
-
-@pytest.mark.asyncio
-async def test_get_host_configuration_no_config(client):
-    """
-    Test that get_host_configuration returns None when site not configured.
-    """
-    from memorial import get_host_configuration
-    
-    with patch.dict("memorial.app.config", {
-        "ARCHIVE_CONFIG": {},
-        "STRIP_PORT": False
-    }, clear=False):
-        normalized_host, config = get_host_configuration("unknown.com")
-        assert config is None
-
-
-@pytest.mark.asyncio
-async def test_get_host_configuration_with_port_stripping(client):
-    """
-    Test that get_host_configuration strips port when STRIP_PORT is enabled.
-    """
-    from memorial import get_host_configuration
-    
-    with patch.dict("memorial.app.config", {
-        "ARCHIVE_CONFIG": {"example.com": {"message_pt": "Test"}},
-        "STRIP_PORT": True
-    }, clear=False):
-        # Test with port
-        normalized_host, config = get_host_configuration("example.com:8080")
-        assert normalized_host == "example.com"
-        assert config is not None
-
-
-
-
-# URL Construction Tests: Verify correct Wayback Machine URLs are constructed
-# Tests: URLs with/without version timestamps, noFrame preference, default URL selection
-# ==============================================================================
-# Tests for URL Construction
-# ==============================================================================
-
-
-@pytest.mark.asyncio
-async def test_wayback_url_with_version(client):
-    """
-    Test that Wayback URLs are correctly constructed with version timestamps.
-    """
-    with with_test_config({
-        "versioned.com": {
-            "version": "20200101120000",
-        }
-    }):
-        response = await request_host(client, "/", "versioned.com", expected_status=200)
-        html_content = (await response.data).decode("utf-8")
-        # URL should contain the version timestamp
-        assert "20200101120000" in html_content
-
-
-@pytest.mark.asyncio
-async def test_wayback_url_without_version(client):
-    """
-    Test that Wayback URLs are constructed without version when not specified.
-    """
-    with with_test_config({"no-version.com": {}}):
-        response = await request_host(client, "/", "no-version.com", expected_status=200)
-        html_content = (await response.data).decode("utf-8")
-        # URL should reference the domain without version prefix
-        assert "arquivo.pt" in html_content
-
-
-@pytest.mark.asyncio
-async def test_noframe_url_preference(client):
-    """
-    Test that noFrame URL is used when link_to_noFrame is True.
-    """
-    with with_test_config({
-        "noframe-pref.com": {
-            "link_to_noFrame": True,
-        }
-    }):
-        response = await request_host(client, "/", "noframe-pref.com", expected_status=200)
-        html_content = (await response.data).decode("utf-8")
-        # Should use noFrame URL
-        assert "noFrame/replay" in html_content
-
-
-@pytest.mark.asyncio
-async def test_regular_wayback_url_used_by_default(client):
-    """
-    Test that regular Wayback URL is used when link_to_noFrame is False or not set.
-    """
-    with with_test_config({
-        "regular-wayback.com": {
-            "link_to_noFrame": False,
-        }
-    }):
-        response = await request_host(client, "/", "regular-wayback.com", expected_status=200)
-        html_content = (await response.data).decode("utf-8")
-        # Should use regular wayback URL
-        assert "wayback" in html_content
-
-
-
-
-# Edge Cases & Error Conditions: Verify robustness with malformed input and edge cases
-# Tests: malformed headers, long paths, unicode, multiple slashes, language content,
-#        custom styling, timeouts, multi-site configurations
-# ==============================================================================
-# Tests for Edge Cases and Error Conditions
-# ==============================================================================
-
-
-@pytest.mark.asyncio
-async def test_malformed_host_header(client):
-    """
-    Test handling of malformed host headers.
-    """
-    with with_test_config({"normal.com": {}}):
-        # Test with empty host header
-        response = await client.get("/", headers={"Host": ""})
-        # Should not crash, might return 502 for unconfigured/invalid host
-        assert response.status_code in [200, 502]
-
-
-@pytest.mark.asyncio
-async def test_very_long_path(client):
-    """
-    Test handling of very long URL paths.
-    """
-    with with_test_config({"long-path.com": {}}):
-        long_path = "/" + "/".join(["segment"] * 50)
-        response = await request_host(client, long_path, "long-path.com", expected_status=200)
-        assert response.status_code == 200
-
-
-@pytest.mark.asyncio
-async def test_unicode_characters_in_path(client):
-    """
-    Test handling of unicode characters in URL path.
-    """
-    with with_test_config({"unicode.com": {}}):
-        response = await request_host(client, "/página/sobre-nós", "unicode.com", expected_status=200)
-        assert response.status_code == 200
-
-
-@pytest.mark.asyncio
-async def test_multiple_slashes_in_path(client):
-    """
-    Test handling of multiple consecutive slashes in path.
-    """
-    with with_test_config({"slashes.com": {}}):
-        response = await request_host(client, "//path///to///resource", "slashes.com", expected_status=200)
-        assert response.status_code == 200
-
-
-@pytest.mark.asyncio
-async def test_language_specific_content(client):
-    """
-    Test that language-specific messages are rendered correctly.
-    """
-    with with_test_config({
-        "i18n-test.com": {
-            "message_pt": "Mensagem em português",
-            "message_en": "Message in English",
-            "default_language": "pt",
-        }
-    }):
-        response = await request_host(client, "/", "i18n-test.com", expected_status=200)
-        html_content = (await response.data).decode("utf-8")
-        # Both messages should be present (one hidden by default)
-        assert "Mensagem em português" in html_content
-        assert "Message in English" in html_content
-
-
-@pytest.mark.asyncio
-async def test_custom_button_color(client):
-    """
-    Test that custom button color is applied.
-    """
-    with with_test_config({
-        "colored-button.com": {
-            "button_color": "#FF5733",
-        }
-    }):
-        response = await request_host(client, "/", "colored-button.com", expected_status=200)
-        html_content = (await response.data).decode("utf-8")
-        assert "#FF5733" in html_content
-
-
-@pytest.mark.asyncio
-async def test_custom_links_for_languages(client):
-    """
-    Test that custom links for different languages are applied.
-    """
-    with with_test_config({
-        "custom-links.com": {
-            "link_pt": "https://example.com/pt",
-            "link_en": "https://example.com/en",
-        }
-    }):
-        response = await request_host(client, "/", "custom-links.com", expected_status=200)
-        html_content = (await response.data).decode("utf-8")
-        assert "https://example.com/pt" in html_content
-        assert "https://example.com/en" in html_content
-
-
-@pytest.mark.asyncio
-async def test_httpx_timeout_configuration(client):
-    """
-    Test that WAYBACK_REQUEST_TIMEOUT configuration is respected.
-    """
-    mock_response = Mock()
-    mock_response.status_code = 200
-    mock_response.headers = {"content-type": "text/html"}
-    mock_response.content = b"<html><head><title>Test</title></head></html>"
-    
-    mock_client_instance = Mock()
-    mock_client_instance.head = AsyncMock(return_value=mock_response)
-    mock_client_instance.get = AsyncMock(return_value=mock_response)
-    mock_client_instance.__aenter__ = AsyncMock(return_value=mock_client_instance)
-    mock_client_instance.__aexit__ = AsyncMock(return_value=None)
-    
-    with patch("memorial.httpx.AsyncClient") as mock_async_client_class:
-        mock_async_client_class.return_value = mock_client_instance
-        
-        with patch.dict("memorial.app.config", {
-            "WAYBACK_REQUEST_TIMEOUT": 10,
-            "EXTRACT_METADATA": True
-        }, clear=False):
-            with with_test_config({"timeout-test.com": {}}):
-                response = await request_host(client, "/", "timeout-test.com", expected_status=200)
-                assert response.status_code == 200
-                # Verify AsyncClient was called with timeout
-                mock_async_client_class.assert_called()
-
-
-@pytest.mark.asyncio
-async def test_multiple_sites_different_configs(client):
-    """
-    Test that multiple sites can be configured with completely different configurations.
-    """
-    with with_test_config({
-        "site-a.com": {
-            "status_code": 200,
-            "message_pt": "Site A",
-            "default_language": "pt",
-            "version": "20190101000000",
-        },
-        "site-b.com": {
-            "status_code": 410,
-            "message_en": "Site B - Gone",
-            "default_language": "en",
-        },
-        "site-c.com": {
-            "status_code": 503,
-            "extract_metadata": True,
-            "version": "20210601000000",
-        },
-    }):
-        # Test site A
-        response_a = await request_host(client, "/", "site-a.com", expected_status=200)
-        assert "Site A" in (await response_a.data).decode("utf-8")
-        
-        # Test site B
-        response_b = await request_host(client, "/", "site-b.com", expected_status=410)
-        assert response_b.status_code == 410
-        
-        # Test site C
-        response_c = await request_host(client, "/", "site-c.com", expected_status=503)
-        assert response_c.status_code == 503
